requirements.txt
absl-py==0.7.1
asn1crypto==0.24.0
astor==0.8.0
attrs==19.1.0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
click-completion==0.5.1
click-didyoumean==0.0.3
click-help-colors==0.6
colorama==0.3.9
comet-git-pure==0.19.11
comet-ml==2.0.8
configobj==5.0.6
cryptography==2.7
cursor==1.3.4
cycler==0.10.0
datadog==0.29.3
decorator==4.4.0
everett==1.0.2
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
gradient==0.2.1
gradient-sdk==0.0.2
gradient-statsd==1.0.1
grpcio==1.22.0
h5py==2.9.0
halo==0.0.28
hyperopt==0.1.2
idna==2.8
image-classifiers==0.2.0
imageio==2.5.0
imagesize==1.1.0
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
log-symbols==0.0.14
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
netifaces==0.10.9
networkx==2.3
numpy==1.17.0
nvidia-ml-py3==7.352.0
Pillow==6.1.0
progressbar2==3.42.0
prompt-toolkit==2.0.9
protobuf==3.9.1
pycparser==2.19
pymongo==3.8.0
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyrsistent==0.15.4
python-dateutil==2.8.0
python-utils==2.3.0
PyWavelets==1.0.3
PyYAML==5.1.2
requests==2.22.0
requests-toolbelt==0.9.1
scikit-image==0.15.0
scikit-learn==0.21.3
scipy==1.3.1
shellingham==1.3.1
six==1.12.0
spinners==0.0.23
termcolor==1.1.0
terminaltables==3.1.0
tqdm==4.33.0
urllib3==1.25.3
wcwidth==0.1.7
websocket-client==0.56.0
Werkzeug==0.15.5
wrapt==1.11.2
wurlitzer==1.0.3
